requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cycler==0.10.0
Django==3.1.4
kiwisolver==1.3.1
matplotlib==3.3.1
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
